kt: Move ciq_helpers.py to ktlib
Useful because kt is exposed as package and these helpers can be used
in multiple places, not only this repo.
