+++ kt/ktlib/ciq_helpers.py
